cast to fix 64bit build warnings. From manpage: POSIX.1-2001. Note that RFC 2553 defines a prototype where the last parameter cnt is of type size_t. Many systems follow RFC 2553. Glibc 2.0 and 2.1 have size_t, but 2.2 has socklen_t.

+++ lib/inet_ntop.h
@@ -20,7 +20,7 @@
  * This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY
  * KIND, either express or implied.
  *
- * $Id: inet_ntop.h,v 1.5 2009-05-18 12:25:45 yangtse Exp $
+ * $Id: inet_ntop.h,v 1.6 2009-08-06 11:10:30 gknauf Exp $
  ***************************************************************************/
 
 #include "setup.h"
@@ -31,7 +31,7 @@ char *Curl_inet_ntop(int af, const void *addr, char *buf, size_t size);
 #ifdef HAVE_ARPA_INET_H
 #include <arpa/inet.h>
 #endif
-#define Curl_inet_ntop(af,addr,buf,size) inet_ntop(af,addr,buf,size)
+#define Curl_inet_ntop(af,addr,buf,size) inet_ntop(af,addr,buf,(size_t)size)
 #endif
 
 #endif /* __INET_NTOP_H */
